+++ backend/src/Command/SendMissingCharacterMail.php
@@ -77,7 +77,7 @@ class SendMissingCharacterMail extends Command
                 's',
                 InputOption::VALUE_OPTIONAL,
                 'Time to sleep in seconds after each mail sent (ESI rate limit is 4/min)',
-                20
+                '20'
             );
         $this->configureLogOutput($this);
     }