+++ backend/src/Command/UpdateCharacters.php
@@ -78,7 +78,7 @@ class UpdateCharacters extends Command
                 's',
                 InputOption::VALUE_OPTIONAL,
                 'Time to sleep in milliseconds after each update',
-                50
+                '50'
             );
         $this->configureLogOutput($this);
     }