+++ backend/src/Command/UpdateCorporations.php
@@ -87,7 +87,7 @@ class UpdateCorporations extends Command
                 's',
                 InputOption::VALUE_OPTIONAL,
                 'Time to sleep in milliseconds after each update',
-                50
+                '50'
             );
         $this->configureLogOutput($this);
     }