+++ backend/src/Command/UpdatePlayerGroups.php
@@ -60,7 +60,7 @@ class UpdatePlayerGroups extends Command
                 's',
                 InputOption::VALUE_OPTIONAL,
                 'Time to sleep in milliseconds after each update',
-                35
+                '35'
             );
         $this->configureLogOutput($this);
     }