+++ backend/src/Command/UpdateServiceAccounts.php
@@ -83,7 +83,7 @@ class UpdateServiceAccounts extends Command
                 's',
                 InputOption::VALUE_OPTIONAL,
                 'Time to sleep in milliseconds after each update',
-                35
+                '35'
             );
         $this->configureLogOutput($this);
     }